Add support for in-memory database (plain object)
The on-disk database may not be desirable for many use cases.
A plain object that can be persisted outside of this module
seems like a good fallback. To allow seamless change from one
to the other, the interfaces are identical, i.e. promises
are used even when they are not strictly necessary.

This commit also adds tests, reusing several existing functions
to verify that the in-memory database operates as expected.